+++ native/src/main/java/io/ballerina/stdlib/sql/Constants.java
@@ -78,6 +78,7 @@ public final class Constants {
 
     public static final String ANNON_RECORD_TYPE_NAME = "record {";
     public static final String BACKTICK = "`";
+    public static final String DEFAULT_STREAM_CONSTRAINT_NAME = "$stream$anon$constraint$";
     /**
      * Constants related connection pool.
      */
@@ -346,7 +347,7 @@ public final class Constants {
         public static final String CIVIL = "time:Civil[]";
         public static final String TIME_OF_DAY = "time:TimeOfDay[]";
         public static final String DECIMAL = "decimal[]";
-        public static final String UTC = "([int,decimal] & readonly)[]";
+        public static final String UTC = "time:Utc[]";
         public static final String BYTE = "byte[][]";
     }