+++ native/src/main/java/io/ballerina/stdlib/sql/nativeimpl/QueryProcessor.java
@@ -25,9 +25,9 @@ import io.ballerina.runtime.api.TypeTags;
 import io.ballerina.runtime.api.creators.TypeCreator;
 import io.ballerina.runtime.api.creators.ValueCreator;
 import io.ballerina.runtime.api.types.RecordType;
-import io.ballerina.runtime.api.types.StructureType;
 import io.ballerina.runtime.api.types.Type;
 import io.ballerina.runtime.api.types.UnionType;
+import io.ballerina.runtime.api.utils.TypeUtils;
 import io.ballerina.runtime.api.values.BError;
 import io.ballerina.runtime.api.values.BMap;
 import io.ballerina.runtime.api.values.BObject;
@@ -121,7 +121,8 @@ public class QueryProcessor {
                 statement = connection.prepareStatement(sqlQuery);
                 statementParameterProcessor.setParams(connection, statement, parameterizedQuery.getInsertions());
                 resultSet = statement.executeQuery();
-                RecordType streamConstraint = (RecordType) ((BTypedesc) recordType).getDescribingType();
+                RecordType streamConstraint = (RecordType) TypeUtils.getReferredType(
+                        ((BTypedesc) recordType).getDescribingType());
                 List<ColumnDefinition> columnDefinitions = Utils.getColumnDefinitions(resultSet, streamConstraint);
                 return ValueCreator.createStreamValue(TypeCreator.createStreamType(streamConstraint,
                         PredefinedTypes.TYPE_NULL), resultParameterProcessor
@@ -175,7 +176,7 @@ public class QueryProcessor {
             AbstractStatementParameterProcessor statementParameterProcessor,
             AbstractResultParameterProcessor resultParameterProcessor, boolean isWithInTrxBlock,
             TransactionResourceManager trxResourceManager) {
-        Type describingType = ballerinaType.getDescribingType();
+        Type describingType = TypeUtils.getReferredType(ballerinaType.getDescribingType());
         Object dbClient = client.getNativeData(Constants.DATABASE_CLIENT);
         if (dbClient != null) {
             SQLDatasource sqlDatasource = (SQLDatasource) dbClient;
@@ -278,10 +279,12 @@ public class QueryProcessor {
     }
 
     public static BMap<BString, Object> createRecord(ResultSet resultSet, List<ColumnDefinition> columnDefinitions,
-                                                     StructureType recordConstraint,
+                                                     RecordType recordConstraint,
                                                      AbstractResultParameterProcessor resultParameterProcessor)
             throws SQLException, DataError {
-        BMap<BString, Object> record = ValueCreator.createMapValue(recordConstraint);
+        BMap<BString, Object> record = ValueCreator.createRecordValue(
+                recordConstraint.getPackage(), recordConstraint.getName()
+        );
         Utils.updateBallerinaRecordFields(resultParameterProcessor, resultSet, record, columnDefinitions);
         return record;
     }