+++ changelog.md
@@ -10,6 +10,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+## [1.4.1] - 2022-06-21
+
+### Changed
+
+- [Fix NullPointerException when retrieving record with default value](https://github.com/ballerina-platform/ballerina-standard-library/issues/2985)
+
 ## [1.4.0] - 2022-05-30
 
 ### Added